Now makefile can track all! (but a nightmare to maintainers)
