+++ src/Grade.h
@@ -7,6 +7,11 @@
 #ifndef GRADE_H
 #define GRADE_H
 
+#include<string>
+#include<sstream>
+#include<fstream>
+#include<iomanip>
+
 /**
  * Grade class.
  *
@@ -24,59 +29,45 @@ public:
    * Default constructor
    *
    */
-  Grade() {
-    _grade = 0.0;
-  }
+  Grade();
 
   /**
    * Parameterised constructor
    *
    * @param  double
    */
-  Grade(double grade) {
-    _grade = grade;
-  }
+  Grade(double);
 
   /**
    * Parameterised constructor
    *
    * @param  const Grade&
    */
-  Grade(const Grade& other) {
-    _grade = other._grade;
-  }
-
-  /**
-   * Overload assign operator
-   *
-   * @param  const Grade&
-   *
-   * @return const Grade&
-   */
-  const Grade& operator=(const Grade& other) {
-    _grade = other._grade;
-    return *this;
-  }
-
-  /**
-   * Overload assign operator
-   *
-   * @param  double
-   *
-   * @return const Grade&
-   */
-  const Grade& operator=(double grade) {
-    _grade = grade;
-    return *this;
-  }
+  Grade(const Grade&);
 
   /**
    * Destructor
    *
    */
-  ~Grade() {
-    // Do nothing.
-  }
+  ~Grade();
+
+  /**
+   * Overload assign operator
+   *
+   * @param  const Grade&
+   *
+   * @return const Grade&
+   */
+  const Grade& operator=(const Grade&);
+
+  /**
+   * Overload assign operator
+   *
+   * @param  double
+   *
+   * @return const Grade&
+   */
+  const Grade& operator=(double);
 
 public:
   /**
@@ -84,148 +75,72 @@ public:
    *
    * @return Grade
    */
-  Grade to4Scale() const {
-    if (_grade < 3.0) return Grade(0.0);
-    if (_grade < 4.0) return Grade(1.0);
-    if (_grade < 5.0) return Grade(1.5);
-    if (_grade < 6.0) return Grade(2.0);
-    if (_grade < 7.0) return Grade(2.5);
-    if (_grade < 8.0) return Grade(3.0);
-    if (_grade < 9.0) return Grade(3.5);
-    return Grade(4.0);
-  }
+  Grade to4Scale() const;
 
   /**
    * Convert grade to A scale.
    *
    * @return std::string
    */
-  std::string toAScale() const {
-    if (_grade < 3.0) return "F";
-    if (_grade < 4.0) return "D";
-    if (_grade < 5.0) return "D+";
-    if (_grade < 6.0) return "C";
-    if (_grade < 7.0) return "B";
-    if (_grade < 8.0) return "B+";
-    if (_grade < 9.0) return "A";
-    return "A+";
-  }
+  std::string toAScale() const;
 
   /**
    * Convert grade to string.
    *
    * @return std::string
    */
-  std::string toString() const {
-    std::stringstream builder;
-
-    builder << std::fixed << std::setprecision(2);
-
-    builder << _grade;
-
-    return builder.str();
-  }
+  std::string toString() const;
 
   /**
    * Arithmetic comparision
    *
    */
 
-  bool operator<(const Grade& other) const {
-    return _grade < other._grade;
-  }
+  bool operator<(const Grade&) const;
 
-  bool operator>(const Grade& other) const {
-    return _grade > other._grade;
-  }
+  bool operator>(const Grade&) const;
 
-  bool operator>=(const Grade& other) const {
-    return _grade >= other._grade;
-  }
+  bool operator>=(const Grade&) const;
 
-  bool operator<=(const Grade& other) const {
-    return _grade <= other._grade;
-  }
+  bool operator<=(const Grade&) const;
 
-  bool operator==(const Grade& other) const {
-    return _grade == other._grade;
-  }
+  bool operator==(const Grade&) const;
 
   /**
    * Arithmetic operators
    *
    */
-  Grade operator+(const Grade& other) const {
-    return Grade(_grade + other._grade);
-  }
+  Grade operator+(const Grade&) const;
 
-  Grade operator+(double other) const {
-    return Grade(_grade + other);
-  }
+  Grade operator+(double) const;
 
-  const Grade& operator+=(const Grade& other) {
-    _grade += other._grade;
-    return *this;
-  }
+  const Grade& operator+=(const Grade&);
 
-  const Grade& operator+=(double other) {
-    _grade += other;
-    return *this;
-  }
+  const Grade& operator+=(double);
 
-  Grade operator-(const Grade& other) const {
-    return Grade(_grade - other._grade);
-  }
+  Grade operator-(const Grade&) const;
 
-  Grade operator-(double other) const {
-    return Grade(_grade - other);
-  }
+  Grade operator-(double) const;
 
-  const Grade& operator-=(const Grade& other) {
-    _grade -= other._grade;
-    return *this;
-  }
+  const Grade& operator-=(const Grade&);
 
-  const Grade& operator-=(double other) {
-    _grade -= other;
-    return *this;
-  }
+  const Grade& operator-=(double);
 
-  Grade operator*(const Grade& other) const {
-    return Grade(_grade * other._grade);
-  }
+  Grade operator*(const Grade&) const;
 
-  Grade operator*(double other) const {
-    return Grade(_grade * other);
-  }
+  Grade operator*(double) const;
 
-  const Grade& operator*=(const Grade& other) {
-    _grade *= other._grade;
-    return *this;
-  }
+  const Grade& operator*=(const Grade&);
 
-  const Grade& operator*=(double other) {
-    _grade *= other;
-    return *this;
-  }
+  const Grade& operator*=(double);
 
-  Grade operator/(const Grade& other) const {
-    return Grade(_grade / other._grade);
-  }
+  Grade operator/(const Grade&) const;
 
-  Grade operator/(double other) const {
-    return Grade(_grade / other);
-  }
+  Grade operator/(double) const;
 
-  const Grade& operator/=(const Grade& other) {
-    _grade /= other._grade;
-    return *this;
-  }
+  const Grade& operator/=(const Grade&);
 
-  const Grade& operator/=(double other) {
-    _grade /= other;
-    return *this;
-  }
+  const Grade& operator/=(double);
 
   /**
    * Cast to ostream.
@@ -235,10 +150,7 @@ public:
    *
    * @return std::ostream&
    */
-  friend std::ostream& operator<<(std::ostream& out, const Grade& grade) {
-    out << grade._grade;
-    return out;
-  }
+  friend std::ostream& operator<<(std::ostream&, const Grade&);
 };
 
 #endif
