--- src/include.h
@@ -1,28 +1,0 @@
-/**
- * Include all necessary files.
- *
- * Code by @trhgquan - https://github.com/trhgquan
- */
-
-#ifndef INCLUDE_H
-#define INCLUDE_H
-
-#include<algorithm>
-#include<iostream>
-#include<iomanip>
-#include<fstream>
-#include<memory>
-#include<sstream>
-#include<string>
-#include<set>
-#include<vector>
-
-// Custom header
-#include"Utility.h"
-#include"CommandLineParser.h"
-#include"InputHelper.h"
-#include"Grade.h"
-#include"Subject.h"
-#include"Personal.h"
-#include"OutputHelper.h"
-#endif
